fix: Complete BigDecimal to Double migration for KMP compatibility
- Migrate Money class from BigDecimal to Double for better KMP support
- Update all Money-related operations (plus, minus, times, div, abs, etc.)
- Fix all usages of BigDecimal.ZERO to 0.0 throughout the codebase
- Update TransactionHistoryState to use Double for percentChange
- Fix CategoryStatsCard to handle Double instead of BigDecimal
- Update MoneyConverter to work with Double serialization
- Fix CompactLayout abs() usage for expense calculations
- Update all use cases to work with Double instead of BigDecimal:
  * CalculateBalanceMetricsUseCase
  * CalculateCategoryStatsUseCase
  * GetCategoriesWithAmountUseCase
  * PredictFutureExpensesUseCase
  * CalculatePeerComparisonUseCase
  * CalculateRetirementForecastUseCase
  * CalculateEnhancedFinancialMetricsUseCase
- Fix all feature modules (statistics, transaction, budget, history, home)
- Ensure full KMP compatibility and successful build
- Remove all BigDecimal dependencies for cleaner KMP code

+++ feature/history/src/main/java/com/davidbugayov/financeanalyzer/presentation/history/state/TransactionHistoryState.kt
@@ -58,7 +58,7 @@ data class TransactionHistoryState(
     val periodType: PeriodType = PeriodType.ALL,
     val startDate: Date = Calendar.getInstance().apply { add(Calendar.YEAR, -5) }.time,
     val endDate: Date = Calendar.getInstance().time,
-    val categoryStats: Triple<Money, Money, BigDecimal?>? = null,
+    val categoryStats: Triple<Money, Money, Double?>? = null,
     val showPeriodDialog: Boolean = false,
     val showCategoryDialog: Boolean = false,
     val showSourceDialog: Boolean = false,
